+++ modules/hanna-react/CHANGELOG.md
@@ -11,6 +11,7 @@
 - feat: Add prop `stacked` to `.CheckboxButtonsGroup` and `.RadioButtonsGroup`
 - feat: Add property `group` to `MultiselectOption` type, for grouping options
 - fix: Apply `wrapperProps.className` to `Heading`'s wrapper element
+- fix: Non-sensical `Gallery` close button label
 
 ## 0.10.109